moved reporting component to Organization component
